TMBF: Increase resolution of detector sweep scale
Turns out that single precision doesn't provide enough resolution.

git-svn-id: svn+ssh://serv2/home/subversion/repos/controls/diamond/trunk/ioc/TMBF@103152 e099a375-04f9-0310-9d5f-a741eaff62e1

+++ tmbfApp/Db/detector.py
@@ -23,7 +23,7 @@ for bunch in range(4):
 
 # This will be reprocessed when necessary, basically when the sequencer settings
 # change and a detector trigger occurs.
-Waveform('DET:SCALE', TUNE_LENGTH, 'FLOAT',
+Waveform('DET:SCALE', TUNE_LENGTH, 'DOUBLE',
     SCAN = 'I/O Intr', DESC = 'Scale for frequency sweep')
 
 # Two overflow detection bits are generated
